# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "service responsible for operations with students and teachers accounts",
        "title": "utf-account-service",
        "contact": {
            "name": "Pyankov Daniil",
            "url": "http://t.me/lifelessdev",
            "email": "pyankovdaniildev@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/teacher": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "create a new teacher",
                "parameters": [
                    {
                        "description": "Required field values to create teacher",
                        "name": "teacherData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/transfer.TeacherCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/transfer.TeacherCreateResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/teacher/verify": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "verify teacher account",
                "parameters": [
                    {
                        "description": "Request body with verifying teacher ID",
                        "name": "teacherID",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/transfer.TeacherVerifyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "definitions": {
        "transfer.TeacherCreateRequest": {
            "type": "object",
            "required": [
                "first_name",
                "last_name",
                "password",
                "report_email",
                "username"
            ],
            "properties": {
                "first_name": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "middle_name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "report_email": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "transfer.TeacherCreateResponse": {
            "type": "object",
            "properties": {
                "teacherID": {
                    "type": "string"
                }
            }
        },
        "transfer.TeacherVerifyRequest": {
            "type": "object",
            "required": [
                "teacher_id"
            ],
            "properties": {
                "teacher_id": {
                    "type": "string"
                }
            }
        }
    }
}